+++ website/src/components/ui/Link.tsx
@@ -1,9 +1,8 @@
 import { Link as RouterLink, LinkProps as RouterLinkProps } from 'react-router-dom'
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export default function Link({ to, className, children, ...props }: RouterLinkProps) {
   return (
-    <RouterLink to={to} className='text-blue-600 rounded-sm bg-slate-100 px-1' {...props}>
+    <RouterLink to={to} className={`text-blue-600 rounded-sm bg-slate-100 px-1 ${className}`} {...props}>
       {children}
     </RouterLink>
   )