+++ packages/react-together/CHANGELOG.md
@@ -9,7 +9,7 @@
 `0.3.0-0`
  - Added support for viewInfo
 
-### 0.2.0
+## 0.2.0
 `2024-11-01`
 
  - Added `useFunctionTogether` hook